+++ src/spellbot/web/builder.py
@@ -74,6 +74,7 @@ def build_web_app() -> web.Application:
             web.post(r"/api/game/{game}/record", rest.game_record_endpoint),
             web.post(r"/api/notification", rest.create_notification_endpoint),
             web.patch(r"/api/notification/{notif}", rest.update_notification_endpoint),
+            web.delete(r"/api/notification/{notif}", rest.delete_notification_endpoint),
         ],
     )
     return app